requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
pyjwt
python-dotenv
alembic
psycopg2